+++ apps/web/src/views/boards/index.tsx
@@ -2,7 +2,6 @@ import { t } from "@lingui/core/macro";
 import { HiArrowDownTray, HiOutlinePlusSmall } from "react-icons/hi2";
 
 import Button from "~/components/Button";
-import FeedbackModal from "~/components/FeedbackModal";
 import Modal from "~/components/modal";
 import { NewWorkspaceForm } from "~/components/NewWorkspaceForm";
 import { PageHead } from "~/components/PageHead";
@@ -69,13 +68,6 @@ export default function BoardsPage({ isTemplate }: { isTemplate?: boolean }) {
         </div>
 
         <>
-          <Modal
-            modalSize="md"
-            isVisible={isOpen && modalContentType === "NEW_FEEDBACK"}
-          >
-            <FeedbackModal />
-          </Modal>
-
           <Modal
             modalSize="sm"
             isVisible={isOpen && modalContentType === "NEW_BOARD"}