+++ apps/web/src/views/settings/AccountSettings.tsx
@@ -2,7 +2,6 @@ import { t } from "@lingui/core/macro";
 import { env } from "next-runtime-env";
 
 import Button from "~/components/Button";
-import FeedbackModal from "~/components/FeedbackModal";
 import Modal from "~/components/modal";
 import { NewWorkspaceForm } from "~/components/NewWorkspaceForm";
 import { PageHead } from "~/components/PageHead";
@@ -106,12 +105,6 @@ export default function AccountSettings() {
       </Modal>
 
       {/* Global modals */}
-      <Modal
-        modalSize="md"
-        isVisible={isOpen && modalContentType === "NEW_FEEDBACK"}
-      >
-        <FeedbackModal />
-      </Modal>
       <Modal
         modalSize="sm"
         isVisible={isOpen && modalContentType === "NEW_WORKSPACE"}